Replace `provider.query()` with `provider.handle_request()`.
`provider.query()` only takes the request's query string but the provider's
suggestions may depend on more than that. For example the upcoming AccuWeather
provider will need to know the geolocation state stored in the request.

This commit leaves the `query()` method defined on the ADM and Wiki Fruit
providers so that the tests can continue to use it. For these providers,
`handle_request()` delegates to `query()`.