[CONSVC-2008] feat: add nav query top pick provider  (#89)
Core logic for handling Top Pick suggestions. This new provider  indexes domains, including possible similar terms that closely match domain (i.e. a typo) from a provided domain list. This provider ensures queries do not start with `http`, are not under the set character limit (currently 4) and match valid criteria . This pull request also includes unit tests and api tests verifying the effectiveness of the provider to provide sample top picks and to also not send invalid suggestions. 

It was determined a trie data structure was not effective or better than using Python dictionaries. The current implementation is simpler and is an effective strategy for efficiently indexing the list of domains to find a match when queried. 

Further enhancements will follow, including some modularization, addition of some future tests, weaving in hierarchy between sponsored suggestions, etc.

Thanks to @ncloudioj for terrific reviews, direction and support and to @pjenvey for additional reviews. 🎉 

Closes #72