docs: add ADR template [do not deploy]
This commit adds a template for our Architecture Decision Records,
using a similar format to Firefox Accounts, and already used for
DISCO-2045.

It's more structured than Nygard's status-context-decision-consequences
template, but explicitly enumerates the alternatives considered and the
pros and cons of each, which has been helpful in other projects (FxA,
Application Services, UniFFI) that use this format.