Add timestamp to index name to prevent collision (#196)
* Improve indexer test code coverage

* Add a timestamp to the index on every run

This will essentially ensure each run of the indexing goes into its own index
(unless we happen to start 2 jobs at the *exact* same time...
which is highly unlikely).

* Add version to CLI options

It defaults to the version in the configuration.

* Change `ensure_index` to `create_index`

This is because we only really care about the index being created.
With the timestamp, we should be fairly certain that we don't have
an existing index. If there exists and existing index, then the job
should fail.

* Add Wikipedia CLI job to Makefile to run locally

This is more of a reference to the command to run this, rather than
actually running the job.