Close index after switching to new index
We are currently running into some issues with heap memory usage,
where unused indice are using up a lot of the memory on the cluster.
This will close the index after switchover so that the index can't be queried,
but also kept on disk so that if we ever need to rollback and recover an index,
we can do that by reopening the index.