Close index after switching to new index (#249)
* Close index after switching to new index

We are currently running into some issues with heap memory usage,
where unused indice are using up a lot of the memory on the cluster.
This will close the index after switchover so that the index can't be queried,
but also kept on disk so that if we ever need to rollback and recover an index,
we can do that by reopening the index.

* Add some documentation on how to recover index

I didn't put in any commands because we currently still don't have
access to Kibana. So the documentation is more conceptual at the moment.

* Update docs/operations/elasticsearch.md

Co-authored-by: Nan Jiang <najiang@mozilla.com>

---------

Co-authored-by: Nan Jiang <najiang@mozilla.com>