Periodically fetch and update AMO information (#304)
* Periodically fetch and update AMO information

The Addons data will be updated on a daily basis via a cron job.
The cron will just iterate through the list of Addons and query
the Addon API and replace the currently cached data.

* Add a cron ticker to retry tasks sooner if failed

This separates the resync interval from the cron retry. We want to
retry the cron more frequently if it is an error.