feat: add score to AMO RS suggestions and fallback scores to chunked RS uploader
Modify the AMO RS uploader job by adding scores to AMO RS suggestions. The
purpose of this is to give AMO RS suggestions a high score to ensure the client
will prefer them to adM suggestions when there are keyword collisions. (There
are currently 22 collisions total.)

This is intended as a quick fix for the upcoming AMO suggestions experiment. The
client already looks for a `score` property in AMO suggestions, but currently
none is defined, so the default 0.2 score is used. If/when AMO suggestions are
enabled by default instead of being enabled only in experiments, we will want to
revisit this.

To implement this in a way that can be used for all RS suggestion types, add a
param to the chunked RS uploader that lets the consumer specify a fallback score
for suggestions. If a suggestion doesn't include a score and a fallback is
defined, the fallback value will be added to the suggestion as its score.