fix: score AMO suggestions below adM suggestions (#349)
Change the score of AMO suggestions so it's less than the score of adM
suggestions. Use the new score for both Merino and remote settings AMO
suggestions. This ensures the client will show sponsored suggestions instead of
AMO suggestions when keyword collisions occur, which according to Product is the
desired behavior for both the addons-suggestion experiment and general release.