feat: add Pocket RS suggestions uploader (#353)
* feat: add Pocket RS suggestions uploader

Add a job that uploads Pocket suggestions to remote settings. The source data is
expected to be in a CSV file because we're collaborating with the Pocket team to
build a list of suggestions and keywords using [this Google sheet](https://docs.google.com/spreadsheets/d/14mjaEYWnHD3VuTT5vq1HCzCz9e98cm8Z5a9fwjtKzGw/edit?usp=sharing).
Running this job requires an initial manual step of exporting the sheet as CSV.
I didn't add the CSV to this commit because I don't think we want to keep it in
the repo.

Move chunked_rs_uploader.py to a new jobs/utils directory so it can be shared
between the AMO and Pocket uploader and any future uploader jobs. Modify it to
zero-pad the chunk start and end indexes in recored IDs similar to how indexes
are zero-padded in adm records.

* add validation to Pocket RS suggestions uploader

* fix a couple problems in the Pocket RS suggestions uploader

* Filter out empty keyword strings before validation
* Convert each `Suggestion` object to a dict when calling
  `uploader.add_suggestion()`