feat: factor out CSV remote settings uploader (#395)
* feat: factor out CSV remote settings uploader

Factor out a generic CSV remote settings uploader job from the Pocket uploader.
We need to upload MDN suggestions and instead of adding another job for that,
let's make a generic uploader job.

We still want validation of the input data. To accomplish that, I added modules
that implement Pydantic base models for Pocket and MDN. Any time we add a new
type of suggestion for remote settings, we can add a new model module.

I updated jobs.md with info on how to run it. Hopefully it also provides a good
explanation of how this all works.

* Add test coverage and remove branches that should never be entered