Add New Tab API that returns Upday Articles (#459)
* Add the basic skeleton for the new tab API.

This currently just returns the locale and language that is passed in the param.

* Add API endpoint to get Upday Recommendations

This adds the Upday connection information and returns the Pocket specific response
to the user. If the Upday provider is not connected, return an empty result
to the user.

* Add some integration tests for New Tab API

* Add tests for the newtab initialization code

* Update documentation on provider fixtures

The provider fixtures for integration tests have been moved to
be only in scope of Firefox Suggest endpoint tests. So, making it
clear in the documentation as well.

* Updated PR comments

* Add additional details to locale and language documentation

Co-authored-by: Nan Jiang <najiang@mozilla.com>

---------

Co-authored-by: Nan Jiang <najiang@mozilla.com>