feat: remove uBlock Origin and a Privacy Badger keyword from AMO suggestions (#463)
This makes three changes:

Remove uBlock Origin from the list of addons, per AK and Product. We actually
removed it from the remote settings dataset a while ago, but I modified the
amo-rs-uploader job in my local merino-py checkout to do that. We should go
ahead and remove it from the source of truth in Merino. Addon suggestions aren't
enabled in Merino, so it doesn't matter too much otherwise.

Remove the "ad tracking" keyword from Privacy Badger, per AK and Product.

Modify amo-rs-uploader so it sorts keywords in the output JSON. (Keywords are
stored as `set`s in `ADDON_KEYWORDS`.) That makes it easier to eyeball changes
to keywords.