[DISCO-2967] deleting records error + CLI flag
Added the `--allow-delete` flag.  If it's not present, don't allow
records to be deleted.  See
https://bugzilla.mozilla.org/show_bug.cgi?id=1908802 for details.  I
turned off the flag for relevancy updates, because the bug only affects
Suggest.

Refactored the CSV uploader so that it records changes it's going to
make and executes them at the end.  This is needed to correctly
determine if there's a deletion. The current pattern is to delete all
records, then update ones from the CSV data -- undoing the delete.