feat: add geonames uploader SNG-1848
Add a `geonames_uploader` job that downloads data from GeoNames, selects a set
of geonames based on the config (like population size), and uploads the selected
geonames to remote settings. This supports city-based weather queries on the
client and lets the client recognize city and state names.

GeoNames is an open-source geographical database of place names worldwide,
including cities, regions, and countries [1]. Technical documentation at [2].

This commit sets the population threshold to 50,000. That results in 1,013
geonames (U.S. cities and states) in the output, which results in 6 records when
using the usual 200-items-per-attachment count. That seems reasonable.

[1] https://www.geonames.org/
[2] https://download.geonames.org/export/dump/readme.txt