Bug 1971234 - Rewrite the geonames uploader to make it simpler to use
This replaces the two `geonames` and `alternates` command with a single `upload`
command. The core job config is defined in the `CONFIGS_BY_COUNTRY` dict.

Another improvement this makes is to download existing RS records only once, at
the start of the job, instead of having both the geonames part and the
alternates part do it. The alternates part also doesn't need to download each
geonames RS attachment anymore. So overall the job is faster to complete.