fix: Ensure that we display most current stored game info
This will use elasticsearch dates in the storage engine, which trickle
out as ISOFormat strings (sigh), but should be more accurate. The reason
is so that when I search, I'm returning results sorted by `Date` DESC (the
event time) and then `Update` DESC (the time the record was last modified).
This will allow me to use the "latest" updated date in the response, in
case there's more than one record that is "in progress".

I also added a "touched" field to the HTTP response, to validate that
the data being presented is within 5 minutes of the last update call.

Issue: DISCO-3828