feat: create `quick_update` job, use `updated` to determine last event
The `quick_update` is a lighter weight score update that does not
process team information, but only updates scores and status for games
that are either in progress or just completed. This is designed to be
run after a `nightly` job which initializes indicies and fetches known
games. This does require the `sports-meta` index which should be created
by now.

Also, switch to using `updated` as the determination for the previous
game to ensure that we are getting the last "updated" score. This should
indicate the prior game in most cases.

NOTE: these fixes are semi-speculative. Observation during events will
be required.

Issue: DISCO-3828